Reorganize modules
Breaks up __init__.py into multiple modules for better
code organization.
