+++ doitlive/ipython_consoles.py
