+++ tests/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from click.testing import CliRunner
+
+import doitlive
+
+@pytest.fixture(scope='session')
+def runner():
+    doitlive.cli.TESTING = True
+    return CliRunner()
